rename createCapTableTypes.ts -> capTableTypes.ts
